refactor: move all new toolbar-related widgets and utilities to appropriate directories
