+++ lib/macos_ui.dart
@@ -19,13 +19,9 @@ export 'src/buttons/pulldown_button.dart';
 export 'src/buttons/push_button.dart';
 export 'src/buttons/radio_button.dart';
 export 'src/buttons/switch.dart';
-export 'src/buttons/toolbar/toolbar_popup.dart';
-export 'src/buttons/toolbar/toolbar_spacer.dart';
 export 'src/buttons/toolbar/toolbar_icon_button.dart';
-export 'src/buttons/toolbar/toolbar_pulldown_button.dart';
 export 'src/buttons/toolbar/toolbar_overflow_button.dart';
-export 'src/buttons/toolbar/toolbar_overflow_menu.dart';
-export 'src/buttons/toolbar/toolbar_overflow_menu_item.dart';
+export 'src/buttons/toolbar/toolbar_pulldown_button.dart';
 export 'src/dialogs/macos_alert_dialog.dart';
 export 'src/fields/text_field.dart';
 export 'src/icon/macos_icon.dart';
@@ -43,7 +39,11 @@ export 'src/layout/scaffold.dart';
 export 'src/layout/sidebar.dart';
 export 'src/layout/sidebar_item.dart';
 export 'src/layout/title_bar.dart';
-export 'src/layout/toolbar.dart';
+export 'src/layout/toolbar/toolbar.dart';
+export 'src/layout/toolbar/toolbar_overflow_menu.dart';
+export 'src/layout/toolbar/toolbar_overflow_menu_item.dart';
+export 'src/layout/toolbar/toolbar_popup.dart';
+export 'src/layout/toolbar/toolbar_spacer.dart';
 export 'src/layout/window.dart';
 export 'src/macos_app.dart';
 export 'src/sheets/macos_sheet.dart';
